+++ server/route/userRoute.js
@@ -1,14 +1,21 @@
 import express from 'express';
 import userController from '../controllers/users';
-import validate from '../middleware/validate';
+import validateUser from '../middleware/validateUser';
 
 const router = express.Router();
 
 router.post(
   '/api/v1/auth/signup',
-  validate.validateUserRegData,
-  validate.validatePassword,
+  validateUser.validateUserRegData,
+  validateUser.validateUserRegPassword,
   userController.registerUser,
+);
+
+router.post(
+  '/api/v1/auth/login',
+  validateUser.validateUserLoginData,
+  validateUser.validateLogin,
+  userController.loginUser,
 );
 
 export default router;
